Refactor workflows into single module
