+++ bioprov/workflows/__init__.py
@@ -5,7 +5,4 @@ __url__ = "https://github.com/vinisalazar/bioprov"
 __version__ = "0.1.21"
 
 
-from .blastn import blastn_alignment
-from .genome_annotation import genome_annotation
-from .kaiju import KaijuWorkflow
-from .wf_parser import WorkflowOptionsParser
+from .workflows import blastn_alignment, genome_annotation, KaijuWorkflow, WorkflowOptionsParser
